# 110-ticketing-microservices/docs/110-ticketing-app-openapi.yaml
openapi: 3.0.0
info:
  description: "In this application, we create Ticketing API microservice that have feature user management, route management, booking management. This ticketing app developed by using stack NodeJS, Express, JWT,
    MongoDB."
  version: 1.0.0
  title: 110-Ticketing App Microservice
...
servers:
  - url: http://localhost:10081
  - url: http://localhost:10082
  - url: http://localhost:10083
paths:
...
  /api/v1/user/create/admin:
    post:
      tags: 
        - user
      summary: Create new user admin, need super admin secret key 
      description: ""
      operationId: "createAdmin"
      parameters:
      - in: header  
        name: secret-key
        required: true 
        description: Your super admin secret-key
        schema:
          type: string
      requestBody:
        content:  
          application/json:
            schema:
              $ref: "#/components/schemas/createNewAdmin"
        description: "User email, user name and password to register"
        required: true
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/createNewUserSuccess"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/createNewUserErrorEmail"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/noSecretKey"
...
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    createNewAdmin:
      type: object
      properties:
        email:
          type: string
          example: john.doe@email.com
        name: 
          type: string
          example: John Doe
        password:
          type: string
          example: 123456
...
    createNewUserSuccess:
      type: object
      properties:
        message:
          type: string
          example: User was registered successfully!
    createNewUserErrorEmail:
      type: object
      properties:
        message:  
          type: string
          example: "user validation failed: email: Please enter a valid email"
...
    noSecretKey:
      type: object
      properties:
        message:  
          type: string
          example: "No secret-key provided!"
...